separacion script login
