+++ index.html
@@ -45,6 +45,7 @@
     </nav>
   </header>
   
+  <script src="scripts/login.js" type="module"></script>
   <script src="scripts/javascript.js" type="module"></script>
 </body>
 </html>
